x86/platform, acpi: Statically assign IRQ numbers in ACPI, hardware reduced mode
We have to be conservative to dynamically assign IRQ number on the platform
in ACPI Hardware-reduced mode. On the Bay Trail-T(ASUS-T100) platform, there
is a RTC device still using the legacy hardcoded IRQ8, which could cause the
following error if RTC is configured:

7486341a98f: genirq: Flags mismatch irq 8. 00000080 (mmc0) vs. 00000000 (rtc0)

So we want to statically assign IRQ numbers in ACPI hardware reduced mode to
fix this error, this also matches with the original IRQ assignment policy.

Signed-off-by: Li Aubrey <aubrey.li@linux.intel.com>
Cc: Alan Cox <alan@linux.intel.com>
Cc: Len Brown <len.brown@intel.com>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Arjan van de Ven <arjan@linux.intel.com>
Signed-off-by: youling257 <youling257@gmail.com>